feature: Save camp leads from repeating themselves
BACKGROUND:
Currently, if you're a camp lead, you need to enter your info in twice.
We don't want you to have to do that, because that's annoying. This
commit allows you to check a box if you're a camp lead, and if that box
is checked, you don't need to enter your contact data again in the form.